GB/AP/MegaDuck/SMS/GG/NES/MSX: suppress cast from integer to pointer of different size warnings with the the new SDCC

+++ gbdk-lib/include/gb/gb.h
@@ -552,7 +552,7 @@ __REG _current_bank;
     @see BANKREF_EXTERN(), BANKREF()
 */
 #ifndef BANK
-#define BANK(VARNAME) ( (uint8_t) & __bank_ ## VARNAME )
+#define BANK(VARNAME) ( (uint16_t) & __bank_ ## VARNAME )
 #endif
 
 /** Creates a reference for retrieving the bank number of a variable or function
